+++ tests/surrealdb/helpers/init.ts
@@ -2,9 +2,9 @@ import { schema } from '../../mocks/schema';
 import { config } from '../mocks/config';
 import { setup } from '../../helpers/setup';
 
-export const init = async () => setup({
+export const init = async () =>
+	setup({
 		config,
 		schema,
-		tqlPathMap: {
-		}
+		tqlPathMap: {},
 	});