+++ tests/test.sh
@@ -13,20 +13,20 @@ done;
 # Create the namespace, database, and user
 cat tests/surrealdb/mocks/database.surql | docker exec -i $CONTAINER_NAME ./surreal sql -u tester -p tester
 # Create the schema
-docker exec -it $CONTAINER_NAME ./surreal import -u tester -p tester --namespace test --database test --endpoint http://localhost:8000 ./tests/surrealdb/mocks/schema.surql
+docker exec -i $CONTAINER_NAME ./surreal import -u tester -p tester --namespace test --database test --endpoint http://localhost:8000 ./tests/surrealdb/mocks/schema.surql
 # Insert data
-docker exec -it $CONTAINER_NAME ./surreal import -u tester -p tester --namespace test --database test --endpoint http://localhost:8000  ./tests/surrealdb/mocks/data.surql
+docker exec -i $CONTAINER_NAME ./surreal import -u tester -p tester --namespace test --database test --endpoint http://localhost:8000 ./tests/surrealdb/mocks/data.surql
 
 # Setup surrealdb database for the multidb test
 # Create the namespace, database, and user
 cat tests/multidb/mocks/database.surql | docker exec -i $CONTAINER_NAME ./surreal sql -u tester -p tester
 # Create the schema
-docker exec -it $CONTAINER_NAME ./surreal import -u tester -p tester --namespace multi_db_test --database test --endpoint http://localhost:8000  ./tests/multidb/mocks/schema.surql
+docker exec -i $CONTAINER_NAME ./surreal import -u tester -p tester --namespace multi_db_test --database test --endpoint http://localhost:8000 ./tests/multidb/mocks/schema.surql
 # Insert data
-docker exec -it $CONTAINER_NAME ./surreal import -u tester -p tester --namespace multi_db_test --database test --endpoint http://localhost:8000  ./tests/multidb/mocks/data.surql
+docker exec -i $CONTAINER_NAME ./surreal import -u tester -p tester --namespace multi_db_test --database test --endpoint http://localhost:8000 ./tests/multidb/mocks/data.surql
 
 # Always stop container, but exit with 1 when tests are failing
-if CONTAINER_NAME=${CONTAINER_NAME} npx vitest run $@;then
+if CONTAINER_NAME=${CONTAINER_NAME} npx vitest run $@; then
     docker stop ${CONTAINER_NAME}
 else
     docker stop ${CONTAINER_NAME} && exit 1